pyproject.toml
[tool.poetry]
name = "coursework-4"
version = "0.1.0"
description = ""
authors = ["u_alex90 <u_alex90@mail.ru>"]
readme = "README.md"
packages = [{include = "coursework_4"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
oyaml = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
